refactor(ui): share connector state utils between social and blockchain connectors
