+++ packages/ui/src/hooks/use-social.ts
@@ -3,8 +3,12 @@ import { useCallback, useContext } from 'react';
 
 import PageContext from '@/Providers/PageContextProvider/PageContext';
 import { getSocialAuthorizationUrl } from '@/apis/interaction';
+import {
+  generateState,
+  storeState,
+  buildSocialLandingUri,
+} from '@/utils/connectors/social-connectors';
 import { getLogtoNativeSdk, isNativeWebview } from '@/utils/native-sdk';
-import { generateState, storeState, buildSocialLandingUri } from '@/utils/social-connectors';
 
 import useApi from './use-api';
 import useErrorHandler from './use-error-handler';
@@ -34,7 +38,7 @@ const useSocial = () => {
       const { id: connectorId } = connector;
 
       const state = generateState();
-      storeState(state, connectorId);
+      storeState(connectorId, state);
 
       const [error, result] = await asyncInvokeSocialSignIn(
         connectorId,
