+++ packages/ui/tsconfig.json
@@ -7,7 +7,7 @@
       "@/*": [
         "./src/*"
       ]
-    },
+    }
   },
   "include": [
     "src",
